Extract top nav into its own component
